added menu layout component
