+++ client/src/app/main-menu/panorama.module.scss
@@ -1,0 +1,35 @@
+.canvas {
+	transition: all 2s ease;
+	animation: fadeIn 2s ease forwards;
+	position: absolute !important;
+	inset: 0;
+	height: 100%;
+	width: 100%;
+	z-index: -1;
+
+	&::after {
+		content: '';
+		display: block;
+		width: 100%;
+		height: 100%;
+		background: radial-gradient(
+			ellipse at center,
+			rgba(0, 0, 0, 0%) 0%,
+			rgba(0, 0, 0, 0%) 70%,
+			rgba(0, 0, 0, 35%) 100%
+		);
+		position: absolute;
+		top: 0;
+		left: 0;
+	}
+}
+
+@keyframes fadeIn {
+	from {
+		opacity: 0;
+	}
+
+	to {
+		opacity: 1;
+	}
+}
