+++ client/src/components/link-container/link-container.tsx
@@ -2,7 +2,7 @@ import css from './link-container.module.scss'
 
 function LinkContainer() {
 	return (
-		<div className={`${css.linkContainer} temp`}>
+		<div className={css.linkContainer}>
 			<a
 				href="https://www.reddit.com/r/HermitCraft/comments/10wksaw/hctcg_online/"
 				target="_blank"
