+++ client/src/components/menu-layout/index.ts
@@ -1,0 +1,1 @@
+export {default} from './menu-layout'
